virtio-rpmb: fix the description for multi-block reads
Previously the text said we fail if block count is set to 1 despite
language elsewhere in the text referring to:

  "For RPMB read request, one virtio buffer including request command
  and the subsequent [block_count] virtio buffers for response data
  are placed in the queue."

and the existence of both max_wr_cnt and max_rd_cnt configuration
variables certainly implying devices should be able to handle
multi-block reads just like writes.

Fix the description as well as format the steps as an enumerated list
to match the style of the previous section describing write handling.

Reported-by: Ruchika Gupta <ruchika.gupta@linaro.org>
Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Reviewed-by: "Huang, Yang" <yang.huang@intel.com>