README.md: convert embedded html to proper MarkDown
There is no point embedding HTML inside the markdown document. It just
makes it messy to edit. The text is unchanged.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

---

GenAI usage: this is a fiddly file to edit by hand even with the full
power of Emacs. However the prompt:

  please rewrite this in github markdown, using ## and ### for
  headings, droping the extraneous <div> and <p> blocks and using
  proper markdown links. Leave the words themselves unchanged.

worked though it like a champ. As the change is mechanical I don't
think there are any copyright concerns so I'm happy to put my s-o-b on
the patch. I'm declaring the usage here because how LLMs can be used
in Open Source project is a live topic of debate.

After the mechanical changes I eyeballed the README on github and
manually tweaked the formatting to match the original layout as
closely as possible.